backend/pyproject.toml
[tool.poetry]
name = "avatar"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
motor = "^3.4.0"
pinecone-client = "^3.2.2"
python-dotenv = "^1.0.1"
pymongo = "^4.6.3"
pydantic = "^2.7.1"
certifi = "^2024.2.2"
python-multipart = "^0.0.9"
tenacity = "^8.2.3"
path = "^16.14.0"
tiktoken = "^0.6.0"
arrow = "^1.3.0"
deepgram-sdk = "^3.2.7"
pyjwt = "^2.8.0"
fuzzywuzzy = "^0.18.0"
openai = "^1.52.0"

[tool.poetry.scripts]
start = "avatar.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
